data_generation/fractal_graph_expansions/requirements.txt
absl-py==0.7.0
numpy==1.16.2
pandas==0.24.2
scikit-image=0.14.2
scikit-learn=0.20.3
scipy==1.2.1
six==1.12.0
tensorflow==1.13.1
